fix(monitor): filter alertquery result according to onecloud resources
1.排除已经删除的资源
2.对已经删除的资源不再持续报警
